client/src/app/components/table/table-pagination/table-pagination.component.ts
import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
import { Game } from 'src/app/models/Game';

@Component({
  selector: 'app-table-pagination',
  templateUrl: './table-pagination.component.html',
  styleUrls: ['./table-pagination.component.css']
})
export class TablePaginationComponent implements OnInit {

  @Input() pageIndex!: number;
  @Input() gamesPages!: Array<Array<Game>>;

  @Output() pageChange: EventEmitter<number> = new EventEmitter<number>();

  constructor() { }

  ngOnInit(): void {
  }

  pageButtonHandler(index: number) {
    this.pageChange.emit(index);
  }

}
